+++ js/main.js
@@ -1,89 +1,4 @@
-const DESCRIPTION = [
-  'На отдыхе',
-  'На работе',
-  'На тренировке',
-  'На улице',
-  'В школе',
-];
+import {createArrayDescriptionsPhoto} from './data.js';
+import './util.js';
 
-const NAME = [
-  'Александр',
-  'Никита',
-  'Иван',
-  'Анастасия',
-  'Надежда',
-  'Дмитрий',
-  'Виктория',
-  'Юлия',
-];
-
-const MESSAGE = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const getRandomArrayElement = (elements) =>
-  elements[getRandomInteger(0, elements.length - 1)];
-
-function createRandomIdFromRangeGenerator(a, b) {
-  const previousValues = [];
-
-  return function () {
-    let currentValue = getRandomInteger(a, b);
-
-    if (previousValues.length >= b - a + 1) {
-      return null;
-    }
-
-    while (previousValues.includes(currentValue)) {
-      currentValue = getRandomInteger(a, b);
-    }
-    previousValues.push(currentValue);
-    return currentValue;
-  };
-}
-
-const generatePhotoId = createRandomIdFromRangeGenerator(1, 25);
-const generateId = createRandomIdFromRangeGenerator(1, 25);
-const generateCommentsId = createRandomIdFromRangeGenerator(1, 999);
-
-const createComments = function () {
-  return {
-    id: generateCommentsId(),
-    avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-    message: getRandomArrayElement(MESSAGE),
-    name: getRandomArrayElement(NAME),
-  };
-};
-
-const createDescriptionPhoto = function () {
-  return {
-    id: generateId(),
-    url: `photos/${generatePhotoId()}.jpg`,
-    description: getRandomArrayElement(DESCRIPTION),
-    likes: getRandomInteger(15, 200),
-    comments: (arrayCommentsPhoto = Array.from(        /* eslint-disable-line */
-      { length: getRandomInteger(1, 16) },
-      createComments
-    )),
-  };
-};
-
-const arrayDescriptionPhoto = Array.from(
-  { length: 25 },
-  createDescriptionPhoto
-);
-
-
-console.log(arrayDescriptionPhoto);    /* eslint-disable-line */ 
+console.log(createArrayDescriptionsPhoto); /* eslint-disable-line */
