+++ js/util.js
@@ -46,3 +46,21 @@ const getAddedString = function (str, minLength, strAdd) {
 };
 
 getAddedString('q', 4, 'we');
+
+// Функция, которая возвращает случайное число из заданного диапазона.
+
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+export {getRandomInteger};
+
+// Функция, которая возвращает случайный элемент из массива.
+
+const getRandomArrayElement = (elements) =>
+  elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomArrayElement};
